mvp_urban_analysis/requirements.txt
# Основные библиотеки для обработки данных
pandas>=1.5.0
numpy>=1.21.0

# Обработка файлов
openpyxl>=3.0.0
xlrd>=2.0.0

# Визуализация
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Веб-интерфейс
flask>=2.3.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0

# Анализ текста и ML
nltk>=3.8.0
gensim>=4.3.0
scikit-learn>=1.1.0
sentence-transformers>=2.2.0

# Кластеризация и снижение размерности
umap-learn>=0.5.0
hdbscan>=0.8.0

# Геокодирование
requests>=2.28.0

# Утилиты
python-dotenv>=0.19.0
tqdm>=4.64.0 